Renames specs
