+++ spec/helpers/syllableCountIteratorSpec.js
