+++ spec/researches/calculateFleschReadingSpec.js
