+++ spec/researches/countSentencesFromDescription.js
