+++ spec/researches/countSentencesFromText.js
