+++ spec/researches/findKeywordInFirstParagraphSpec.js
