+++ spec/researches/findTransitionWordsSpec.js
