+++ spec/researches/getKeywordDensitySpec.js
