+++ spec/researches/getLinkStatisticsSpec.js
