+++ spec/researches/getPassiveVoiceSpec.js
