+++ spec/researches/imageAltTagsSpec.js
