+++ spec/researches/imageCountInTextSpec.js
