+++ spec/researches/stopWordsInTextSpecs.js
