+++ spec/researches/stopWordsInUrlSpec.js
