+++ spec/stringProcessing/checkNofollowSpec.js
