+++ spec/stringProcessing/countSentencesSpec.js
