+++ spec/stringProcessing/countWordsSpec.js
