+++ spec/stringProcessing/createRegexFromArray.js
