+++ spec/stringProcessing/createRegexFromDoubleArraySpec.js
