+++ spec/stringProcessing/createWordRegexSpec.js
