+++ spec/stringProcessing/findKeywordInUrlSpec.js
@@ -5,3 +5,12 @@ describe( "findKeywordInUrl", function() {
 		expect( findKeywordInUrl( "bogus" ) ).toBe( false );
 	});
 });
+
+describe( "checks keyword occurences in links", function() {
+	it("returns keywords found", function () {
+		expect(findKeywordInUrl( "<a href='http://yoast.com'>test</a>", "yoast" ) ).toBe( false );
+		expect(findKeywordInUrl( "<a href='http://yoast.com'>yoast</a>", "yoast" ) ).toBe( true );
+		expect(findKeywordInUrl( "<a href='http://yoast.com'>$yoast</a>", "$yoast" ) ).toBe( true );
+	});
+});
+
