+++ spec/stringProcessing/getAlttagContent.js
