+++ spec/stringProcessing/getAnchorsFromText.js
