--- spec/stringProcessing/linkKeywordSpec.js
@@ -1,10 +1,0 @@
-var linkKeywordFunction = require( "../../js/stringProcessing/findKeywordInUrl.js" );
-
-describe( "checks keyword occurences in links", function() {
-	it("returns keywords found", function () {
-		expect(linkKeywordFunction( "<a href='http://yoast.com'>test</a>", "yoast" ) ).toBe( false );
-		expect(linkKeywordFunction( "<a href='http://yoast.com'>yoast</a>", "yoast" ) ).toBe( true );
-		expect(linkKeywordFunction( "<a href='http://yoast.com'>$yoast</a>", "$yoast" ) ).toBe( true );
-	});
-});
-
