+++ spec/stringProcessing/matchStringWithRegexSpec.js
