+++ spec/stringProcessing/matchTextWithArraySpec.js
