+++ spec/stringProcessing/matchTextWithWordSpec.js
