+++ spec/stringProcessing/removeNonWordCharacterSpec.js
