+++ spec/stringProcessing/replaceDiacriticsSpecs.js
