+++ spec/stringProcessing/stripNonTextTagsSpec.js
